Resolve WoWUnit - npc_flag is now uint64_t instead of uint32_t
Related to #769 - WIP

+++ src/world/Data/WoWUnit.hpp
@@ -485,7 +485,7 @@ struct WoWUnit : WoWObject
     float mod_ranged_haste;
     float mod_haste_regen;
     uint32_t created_by_spell_id;
-    uint32_t npc_flags;
+    uint64_t npc_flags;
     uint32_t npc_emote_state;
     uint32_t stat[WOWUNIT_STAT_COUNT];
     uint32_t positive_stat[WOWUNIT_STAT_COUNT];
